+++ utils/banner.js
@@ -5,14 +5,12 @@ const Package = require('../package.json');
 
 // Banner function
 function banner() {
-  const lines = [
+  return [
     `Lime version ${Package.version}`,
     Package.description,
     `2021 ${Package.author}, ${Package.license} license`,
     Package.homepage.slice(0, -7),
-  ];
-
-  return lines.join('\n');
+  ].join('\n');
 }
 
 /* ------------------------ division ------------------------ */